requirements.txt
# Minimum packages for azure functions
azure-functions
azure-functions-worker
grpcio
grpcio-tools
protobuf
six
# Additional packages
requests
feedparser
Pillow
numpy
onnxruntime